Move census protocol to model project
