+++ client/src/test/scala/grasshopper/client/census/protocol/CensusJsonProtocolSpec.scala
@@ -1,25 +1,11 @@
 package grasshopper.client.census.protocol
 
 import grasshopper.model.Status
+import grasshopper.protocol.StatusJsonProtocol
 import org.scalatest.{ MustMatchers, FlatSpec }
 import spray.json._
 
-class CensusJsonProtocolSpec extends FlatSpec with MustMatchers with CensusJsonProtocol {
-  "A CensusStatus" must "deserialize from JSON" in {
-    val statusStr = """
-    {
-      "host": "yourhost.local",
-      "status": "OK",
-      "time": "2015-05-06T19:14:19.304850+00:00",
-      "service": "grasshopper-census"
-    }
-                    """
-    val addressStatus = statusStr.parseJson.convertTo[Status]
-    addressStatus.host mustBe "yourhost.local"
-    addressStatus.status mustBe "OK"
-    addressStatus.time mustBe "2015-05-06T19:14:19.304850+00:00"
-    addressStatus.service mustBe "grasshopper-census"
-  }
+class CensusJsonProtocolSpec extends FlatSpec with MustMatchers with StatusJsonProtocol {
 
   it must "serialize to JSON" in {
     val parserStatus = Status(
